small refactoring and bug fix
